api/requirements.txt
fastapi==0.115.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
uvicorn==0.34.0
